Correcciones en namespaces de la capa de servicio.
