+++ mvc/models/CapaServicio/ViajesModelo.php
@@ -1,12 +1,11 @@
 <?php
 namespace app\models\CapaServicio;
-include('ServicioBD/Operaciones.php');
 use Yii;
 use yii\base\Model;
+use app\models\CapaServicio\ServicioBD\GetInfo;
 use app\models\CapaServicio\ServicioBD\Alta;
 use app\models\CapaServicio\ServicioBD\Baja;
 use app\models\CapaServicio\ServicioBD\Modificacion;
-use app\models\CapaServicio\ServicioBD\GetInfo;
 
 /**
  * ViajesModelo short summary.
