+++ mvc/views/agencia/listaViajesTotales.php
@@ -4,8 +4,11 @@ use yii\helpers\BaseHtml;
 use yii\helpers\Html;
 use yii\widgets\ActiveForm;
 use app\assets\PSCssAsset;
+use yii\grid\GridView;
 
 PSCssAsset::register($this);
+
+
 ?>
 <!--<div class="container">
     <section id="main">
@@ -14,105 +17,21 @@ PSCssAsset::register($this);
 <script async defer
 src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDMVbdR-TGis783bW9rB9tZUJXVXsIRzkQ&libraries=places"></script>
 <div class="site-contact">
-    <section id="main">
-        <article>
-            <div id="page-single-main">
-                <br />
-                <h1 id="title-form">
-                    <strong>Listado de Viajes Turno Totales</strong>
-                </h1>
-                <div class="container-form" id="contenedor-formulario">
-                    <h1>
-                        <?= Html::encode($this->title) ?>
-                    </h1>
+    <?= GridView::widget([
+'dataProvider' => $model->dataProvider,
+'columns' => [
+        'ClienteNombre',
+        'OrigenDireccion',
+        'DestinoDireccion',
+        'ChoferNombre',
+        'VehiculoMarca',
+        'VehiculoModelo',
+        'FechaSalida',
+        'ImporteTotal',
+        'Distancia',
+        'ViajeTipo',
+        'Estado',
 
-                    <?php if (Yii::$app->session->hasFlash('Usuario creado con exito')): ?>
-                        <div class="alert alert-success">
-                            Thank you for contacting us. We will respond to you as soon as possible.
-                        </div>
-                        <p>
-                            Note that if you turn on the Yii debugger, you should be able
-                            to view the mail message on the mail panel of the debugger.
-                            <?php if (Yii::$app->mailer->useFileTransport): ?>
-                                Because the application is in development mode, the email is not sent but saved as
-                                a file under
-                                <code>
-                                    <?= Yii::getAlias(Yii::$app->mailer->fileTransportPath) ?>
-                                </code>.
-                                Please configure the
-                                <code>useFileTransport</code>property of the
-                                <code>mail</code>
-                                application component to be false to enable email sending.
-                            <?php endif; ?>
-                        </p>
-                    <?php else: ?>
-                        <div>
-                            <h1>Ac&aacute; colocamos la grilla para listar veh&iacute;culos</h1>
+    ],]); ?>
 
-                            <table class="table table-striped table-hover ">
-                                <thead>
-                                    <tr>
-                                        <th>#</th>
-                                        <th>Column heading</th>
-                                        <th>Column heading</th>
-                                        <th>Column heading</th>
-                                    </tr>
-                                </thead>
-                                <tbody>
-                                    <tr>
-                                        <td>1</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                    </tr>
-                                    <tr>
-                                        <td>2</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                    </tr>
-                                    <tr class="info">
-                                        <td>3</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                    </tr>
-                                    <tr class="success">
-                                        <td>4</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                    </tr>
-                                    <tr class="danger">
-                                        <td>5</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                    </tr>
-                                    <tr class="warning">
-                                        <td>6</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                    </tr>
-                                    <tr class="active">
-                                        <td>7</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                        <td>Column content</td>
-                                    </tr>
-                                </tbody>
-                            </table> 
-                        </div>
-                        <div id='botones-group'>
-                            <?= Html::submitButton('Ver', ['class' => 'btn btn-primary', 'id' => 'btn-guardar']); ?>
-                            &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                            <?= Html::submitButton('Eliminar', ['class' => 'btn btn-primary', 'id' => 'btn-guardar']); ?>
-                            &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                            <?= Html::button('Cancelar', ['class' => 'btn btn-primary', 'id' => 'btn-cancelar']); ?>
-                        </div>
-                    <?php endif; ?>
-                </div>  
-        </article>
-    </section>
 </div>
